pyproject.toml
[project]
name = "vision-llm-web-agent"
version = "0.1.0"
description = "Autonomous web agent powered by Vision Language Models and Playwright"
authors = [
    { name = "Yueheng Zeng", email = "yueheng.zeng@outlook.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.3.0",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vision_llm_web_agent"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
